ventas_experimento/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
google-cloud-pubsub
